Method `Multi::zipStrict()` renamed to `Multi::zipEqual()`.
